PartialConvGan/requirements.txt
torch>=1.12
torchvision>=0.13
numpy
opencv-python
Pillow
tqdm
matplotlib
scikit-image
scipy
